requirements.txt
asgiref==3.8.1
astroid==3.3.10
bcrypt==4.3.0
colorama==0.4.6
dill==0.4.0
Django==5.2.3
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
isort==6.0.1
mccabe==0.7.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
